--- src/components/Market/Detail/SellerInfoTable/MemberInfoTable/MemberInfoTableForm.styled.ts
@@ -1,16 +1,0 @@
-import styled from "styled-components";
-import { SellerInfoHeader, SellerInfoContent } from "../SellerInfoTableForm.styled";
-
-export const MemberInfoWrapper = styled.div`
-    width: 780px;
-    height : 373px;
-`;
-
-export const MemeberTable = styled.table`
-    width: 100%;
-    margin-top : 20px;
-`;
-
-export const MemberHeader = styled( SellerInfoHeader )``;
-
-export const MemberContent= styled( SellerInfoContent )``;
