--- src/components/Market/Detail/SubTitleTable/SubTitleTableForm.styeld.ts
@@ -1,8 +1,0 @@
-import styled from "styled-components";
-import { ContentWrapperAttrs, TableAttrs } from "../../ForSaleListDetail/ForSaleListDetailForm.styled";
-
-export const SubTitleForSaleIntoTableWrapper = styled( ContentWrapperAttrs )`
-    height : 53px;
-`;
-
-export const SubTitleTable = styled( TableAttrs )``;
