--- src/components/Market/Detail/VehicleDetailTable/VehicleDetailTableForm.styled.ts
@@ -1,9 +1,0 @@
-import styled from "styled-components";
-import { ContentWrapperAttrs, TableAttrs } from "../../ForSaleListDetail/ForSaleListDetailForm.styled";
-
-// 차량 설명글
-export const VehicleDetailForSaleInfoTableWrapper = styled( ContentWrapperAttrs )`
-    height: 116px;
-`;
-
-export const VehicleDetailForSaleInfoTable = styled( TableAttrs )``;
