--- src/components/Market/ReadMe.txt
@@ -1,3 +1,0 @@
-Author      : 지용
-Date        : 2022-12-17
-Comments    : Market 디렉토리 설명
