--- src/components/Market/VehicleRegistrationInquiry/PageItems/TopBanner/TopBannerForm.styled.ts
@@ -1,22 +1,0 @@
-import styled from "styled-components";
-import { Button } from "../../../../Common/Button/ButtonForm.styled";
-
-export const TopBannerWrapper = styled.div`
-    display: flex;
-    justify-content: space-between;
-    align-items: center;
-
-    height: 44px;
-    margin-bottom : 20px;
-`;
-
-export const TopBannerLeftWrapper = styled.div`
-    display: flex;
-    justify-content: space-between;
-    align-items: center;
-
-    width: 514px;
-    height : 100%;
-`;
-
-export const TopBannerButton = styled( Button )``;
