--- src/components/Market/VehicleRegistrationInquiry/VehicleRegistrationInquriyForm.styled.ts
@@ -1,8 +1,0 @@
-import styled from "styled-components"
-
-export const Wrapper = styled.div`
-    display: flex;
-    flex-direction: column;
-    width : 100%;
-    padding : 40px 40px;
-`;
