+++ src/types/CommunityType.ts
@@ -1,0 +1,5 @@
+export interface CommunityPropsType {
+    offset : number;
+    postsPerPage : number;
+    totalContentsCount : number;
+};
